base exposion functionality
