+++ SpaceShip.hpp
@@ -20,7 +20,7 @@
 #define ROTATION_DELAY 60
 #define DISPLAY_DELAY 10
 #define SHOOTING_DELAY 100
-#define CHANGE_POSITION_DELAY 50
+#define CHANGE_POSITION_DELAY 30
 #define INERTIA_DELAY 10
 #define INERTIA_COUNTER 500
 #define PROJ_LIFETIME 5000
