+++ common.hpp
@@ -36,6 +36,27 @@ public:
     }
 };
 
+struct DirectionXY{
+    double x;
+    double y;
+    DirectionXY(double, double);
+    DirectionXY();
+
+    DirectionXY& operator *=(int value){
+        x *= value;
+        y *= value;
+        return *this;
+    }
+    DirectionXY(const DirectionXY& tmpXY) : x(tmpXY.x), y(tmpXY.y){}
+};
+
+struct DirectionalVector{
+    DirectionXY p1, p2;
+    DirectionalVector();
+    DirectionalVector(DirectionXY, DirectionXY);
+};
+
+
 Point get_rotated_point(Point, Point, double angle=0.0);
 class Common {
 public:
